lib: media_player: add stop/set_media async support
This commit adds 2 new function into 3.0.12:
 - libvlc_media_player_stop_async()
 - libvlc_media_player_set_media_async()

These 2 new functions won't wait for the input_thread termination. This
will allow LibVLC users to set media and stop them from the mainthread
without blocking. In that case (*_async() function is called), the input
thread will be terminated from a background thread: the
destructor_thread.

libvlc_media_player_release() will wait for this destructor thread
termination. That is why, it is advised to release the media_player at
the end of the program (when UI is destroyed).

This functionality is backport of the async stop functionality of the
new VLC 4.0 player.